+++ src/main/java/com/gas/app/config/RedisConfig.java
@@ -13,7 +13,7 @@ import org.springframework.data.redis.serializer.RedisSerializationContext;
 import java.time.Duration;
 
 @Configuration
-public class RedisConfiguration {
+public class RedisConfig {
     @Value("${spring.redis.host}")
     private String redisHost;
