+++ src/main/java/com/gas/app/controller/api/general/FeedbackRestController.java
@@ -2,7 +2,7 @@ package com.gas.app.controller.api.general;
 
 import com.gas.app.dto.feedback.FeedbackDto;
 import com.gas.app.entity.general.Feedback;
-import com.gas.app.service.general.FeedbackService;
+import com.gas.app.service.feedback.FeedbackService;
 import com.gas.app.util.ResponseHandler;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
