+++ src/main/java/com/gas/app/repository/telegram/TelegramUserPersonalGasAccountRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.jpa.repository.Query;
 import java.util.List;
 import java.util.Optional;
 
-public interface TelegramUserGasPersonalAccountRepository extends JpaRepository<TelegramUserGasPersonalAccount,
+public interface TelegramUserPersonalGasAccountRepository extends JpaRepository<TelegramUserGasPersonalAccount,
         TelegramUserGasPersonalAccountKey> {
 
     @Query("select distinct telegramUserPersonalGasAccount " +
